Move package to ltd.nextalone.*
