+++ app/proguard-rules.pro
@@ -6,6 +6,9 @@
 -keep class org.** { *; }
 -keep class com.microsoft.** { *; }
 -keep class com.rymmmmm.** { *; }
+-keep class cn.lliiooll.** { *; }
+-keep class ltd.nextalone.** { *; }
+-keep class cc.ioctl.** { *; }
 
 -keep public class * extends android.app.Activity
 -keep public class * extends android.app.Application
