+++ app/src/main/java/ltd/nextalone/hook/RemoveIntimateDrawer.kt
@@ -16,30 +16,23 @@
  * along with this software.  If not, see
  * <https://www.gnu.org/licenses/>.
  */
-package me.nextalone.hook
+package ltd.nextalone.hook
 
-import me.nextalone.util.hookBefore
-import me.nextalone.util.hookFalse
-import me.nextalone.util.methods
-import me.singleneuron.qn_kernel.data.hostInfo
-import me.singleneuron.qn_kernel.tlb.ConfigTable
-import me.singleneuron.util.QQVersion
+import android.view.View
+import ltd.nextalone.util.hookBefore
+import ltd.nextalone.util.hookNull
 import nil.nadph.qnotified.hook.CommonDelayableHook
+import nil.nadph.qnotified.util.DexKit
 import nil.nadph.qnotified.util.Utils
 import java.lang.reflect.Method
 
-object HideTotalNumber : CommonDelayableHook("na_hide_total_number") {
+object RemoveIntimateDrawer : CommonDelayableHook("kr_remove_intimate_drawer") {
 
     override fun initOnce(): Boolean {
         return try {
-            var className = "com.tencent.mobileqq.activity.aio.core.TroopChatPie"
-            if (hostInfo.versionCode <= QQVersion.QQ_8_4_8) {
-                className = "com.tencent.mobileqq.activity.aio.rebuild.TroopChatPie"
-            }
-            for (m: Method in className.methods) {
-                val argt = m.parameterTypes
-                if (m.name == ConfigTable.getConfig(HideTotalNumber::class.simpleName) && argt.isEmpty()) {
-                    m.hookBefore(this, hookFalse)
+            for (m: Method in DexKit.doFindClass(DexKit.C_IntimateDrawer)!!.declaredMethods) {
+                if (m.name == "a" && m.returnType == View::class.java) {
+                    m.hookBefore(this, hookNull)
                 }
             }
             true
