+++ app/src/main/java/nil/nadph/qnotified/activity/SettingsActivity.java
@@ -22,7 +22,6 @@ import android.annotation.SuppressLint;
 import android.app.AlertDialog;
 import android.content.Context;
 import android.content.DialogInterface;
-import android.content.pm.PackageManager;
 import android.content.res.ColorStateList;
 import android.content.res.Resources;
 import android.graphics.Color;
@@ -71,9 +70,9 @@ import me.ketal.hook.SendFavoriteHook;
 import me.ketal.hook.SortAtPanel;
 import me.kyuubiran.hook.AutoMosaicName;
 import me.kyuubiran.hook.ShowSelfMsgByLeft;
-import me.nextalone.hook.SimplifyChatLongItem;
-import me.nextalone.hook.SimplifyPlusPanel;
-import me.nextalone.hook.SimplifyQQSettings;
+import ltd.nextalone.hook.SimplifyChatLongItem;
+import ltd.nextalone.hook.SimplifyPlusPanel;
+import ltd.nextalone.hook.SimplifyQQSettings;
 import me.singleneuron.activity.ChangeDrawerWidthActivity;
 import me.singleneuron.hook.*;
 import me.singleneuron.hook.decorator.DisableQzoneSlideCamera;
