+++ app/src/main/java/nil/nadph/qnotified/hook/AbsDelayableHook.java
@@ -93,17 +93,17 @@ import me.kyuubiran.hook.RevokeMsg;
 import me.kyuubiran.hook.ShowSelfMsgByLeft;
 import me.kyuubiran.hook.SimplifyQQSettingMe;
 import me.kyuubiran.hook.testhook.CutMessage;
-import me.nextalone.hook.EnableQLog;
-import me.nextalone.hook.ForcedSendOriginalPhoto;
-import me.nextalone.hook.HideChatVipImage;
-import me.nextalone.hook.HideOnlineNumber;
-import me.nextalone.hook.HideOnlineStatus;
-import me.nextalone.hook.HideProfileBubble;
-import me.nextalone.hook.HideTotalNumber;
-import me.nextalone.hook.RemoveIntimateDrawer;
-import me.nextalone.hook.SimplifyChatLongItem;
-import me.nextalone.hook.SimplifyPlusPanel;
-import me.nextalone.hook.SimplifyQQSettings;
+import ltd.nextalone.hook.EnableQLog;
+import ltd.nextalone.hook.ForcedSendOriginalPhoto;
+import ltd.nextalone.hook.HideChatVipImage;
+import ltd.nextalone.hook.HideOnlineNumber;
+import ltd.nextalone.hook.HideOnlineStatus;
+import ltd.nextalone.hook.HideProfileBubble;
+import ltd.nextalone.hook.HideTotalNumber;
+import ltd.nextalone.hook.RemoveIntimateDrawer;
+import ltd.nextalone.hook.SimplifyChatLongItem;
+import ltd.nextalone.hook.SimplifyPlusPanel;
+import ltd.nextalone.hook.SimplifyQQSettings;
 import me.singleneuron.hook.ChangeDrawerWidth;
 import me.singleneuron.hook.DebugDump;
 import me.singleneuron.hook.ForceSystemAlbum;
